backend/requirements.txt
Flask
Flask-Cors
PyJWT
python-dotenv

